PIC Microcontroller GPS Pointer
This project was created using a template that we used as a starting point for all homeworks and projects the entire semester written by our professor. All code beyond this template was written by my partner and I for the purpose of creating mechanism that when a GPS coordinate is given a vector is calculated between the location of the pointer and a location. This was intented to be used in conjunction with Karsen's senior project for making an high gain directional antenna that would track a UAV through its mission. This code was ment to work by aquiring GPS data from a sensor but the provided microchip was malfunctioning so we reverted to just using the coordinates of the lab we were working from.