Create RGB LED Controller
This code was written in the Arduino IDE and was originally intended for use with a custom RGB LED strip connected to an ATMEGA328P that I made for the gas tank of my motorcycle.  The program has two modes, which users can switch between using a push button.  The first cycles through various colors automatically while the user controls the speed at which the program cycles using a 10K ohm potentiometer.  The second allows the user to choose a color using the potentiometer to manually cycle through colors